+++ SQLiteStudio3/guiSQLiteStudio/dialogs/populatedialog.cpp
@@ -84,6 +84,8 @@ void PopulateDialog::init()
 
     connect(ui->databaseCombo, SIGNAL(currentTextChanged(QString)), this, SLOT(refreshTables()));
     connect(ui->tableCombo, SIGNAL(currentTextChanged(QString)), this, SLOT(refreshColumns()));
+    connect(ui->bottomSelAllBtn, SIGNAL(clicked(bool)), this, SLOT(selectAll()));
+    connect(ui->bottomDeselAllBtn, SIGNAL(clicked(bool)), this, SLOT(deselectAll()));
     connect(POPULATE_MANAGER, SIGNAL(populatingFinished()), widgetCover, SLOT(hide()));
     connect(POPULATE_MANAGER, SIGNAL(finishedStep(int)), widgetCover, SLOT(setProgress(int)));
     connect(POPULATE_MANAGER, SIGNAL(populatingSuccessful()), this, SLOT(finished()));
@@ -184,7 +186,7 @@ void PopulateDialog::refreshColumns()
         ui->rowsSpin->setValue(rows);
 
     int row = 0;
-    for (const QString& column : columns)
+    for (QString& column : columns)
     {
         check = new QCheckBox(column);
         if (columnConfig.contains(column))
@@ -346,6 +348,18 @@ void PopulateDialog::updateState()
 void PopulateDialog::finished()
 {
     QDialog::accept();
+}
+
+void PopulateDialog::selectAll()
+{
+    for (auto&& entry : columnEntries)
+        entry.check->setCheckState(Qt::Checked);
+}
+
+void PopulateDialog::deselectAll()
+{
+    for (auto&& entry : columnEntries)
+        entry.check->setCheckState(Qt::Unchecked);
 }
 
 void PopulateDialog::accept()